feat: WPF GUI refactor (Views, FeatureService split, deep-dive fixes)
- WPF: XAML in UserControls aufgeteilt (SortView, SettingsView, ResultView, ToolsView)
- FeatureService + FeatureCommandService in partial classes gesplittet
- Services: ISettingsService, IThemeService Interfaces, SettingsDto
- MainViewModel: Validation (INotifyDataErrorInfo), RunPipeline, Filters, Settings
- Deep-Dive v1 (Kriterien 1-8): 26/33 Fixes
- Deep-Dive v2 (Kriterien 9-22): 8/12 High-Prio Fixes
  - V2-H03: hashCache thread-safe (ConcurrentDictionary)
  - V2-H05: API SSE heartbeat (15s)
  - V2-H06: SettingsLoader enum validation
  - V2-H07: MoveItemSafely TOCTOU elimination
  - V2-H09: Watch-mode 30s cooldown
  - V2-M03: WPF AddLog shutdown safety
  - V2-M10: API mode case-insensitive
  - V2-M25: AsciiFold ligatures (AE, OE, etc.)
- Neue Tests: CoverageBoost Phase 2-9, Converter, FCS, WPF
- Plan-Dokumente archiviert
- 2883 Tests bestanden, 0 Fehler

+++ src/RomCleanup.CLI/Program.cs
@@ -298,7 +298,7 @@ internal static class Program
                 SizeBytes = group.Winner.SizeBytes,
                 RegionScore = group.Winner.RegionScore,
                 FormatScore = group.Winner.FormatScore,
-                VersionScore = group.Winner.VersionScore,
+                VersionScore = (int)group.Winner.VersionScore,
                 Console = group.Winner.ConsoleKey ?? "",
                 DatMatch = group.Winner.DatMatch
             });
@@ -317,7 +317,7 @@ internal static class Program
                     SizeBytes = loser.SizeBytes,
                     RegionScore = loser.RegionScore,
                     FormatScore = loser.FormatScore,
-                    VersionScore = loser.VersionScore,
+                    VersionScore = (int)loser.VersionScore,
                     Console = loser.ConsoleKey ?? "",
                     DatMatch = loser.DatMatch
                 });
